feature : 일반공격 강화 3회사 구현

+++ Assets/02_Scripts/00_Constants/Constants.cs
@@ -129,7 +129,7 @@ public static class Constants
     /// <summary>
     ///  드론 공격력
     /// </summary>
-    public static int DRONE_ATTACK = 2;
+    public static float DRONE_ATTACK = 2;
 
 
     /// <summary>
